refactor: phase 10 - finalize modular architecture with blueprints, services, and socket separation
