+++ src/routes/api.py
@@ -1,8 +1,16 @@
-from flask import Blueprint, request, jsonify, session
+import socket
+import ssl
+
+import requests
+from flask import Blueprint, request, jsonify, session, send_file
 
 from services.ai_service import generate_port_analysis, generate_db_analysis
 from core.mapper import TopologyMapper
+from core.osint_engine import OSINTEngine
+from core.whois_lookup import WhoisLookup
 from extensions import logger
+from services.report_service import export_scan_report, generate_ai_report
+from services.storage_service import save_history
 
 
 api_bp = Blueprint('api', __name__)
@@ -16,7 +24,7 @@ def save_settings_api():
         session['scanner_settings'] = settings
         return jsonify({'status': 'success', 'message': 'Settings saved'}), 200
     except Exception as e:
-        logger.error(f"Error saving settings: {e}")
+        logger.exception("Error saving settings")
         return jsonify({'status': 'error', 'message': str(e)}), 500
 
 
@@ -27,7 +35,7 @@ def get_settings_api():
         settings = session.get('scanner_settings', {})
         return jsonify(settings), 200
     except Exception as e:
-        logger.error(f"Error retrieving settings: {e}")
+        logger.exception("Error retrieving settings")
         return jsonify({'status': 'error', 'message': str(e)}), 500
 
 
@@ -70,3 +78,204 @@ def db_analysis():
     )
     status_code = result.pop('_status_code', 200)
     return jsonify(result), status_code
+
+
+@api_bp.route('/clear-history', methods=['POST'])
+def clear_history():
+    """Clear all scan history"""
+    try:
+        save_history([])
+        return jsonify({'status': 'success', 'message': 'All scan history cleared successfully'})
+    except Exception as e:
+        logger.exception("Error clearing history")
+        return jsonify({'status': 'error', 'message': str(e)}), 500
+
+
+@api_bp.route('/export/<scan_id>', methods=['GET'])
+def export_report(scan_id):
+    result = export_scan_report(scan_id)
+    if not result["success"]:
+        return result["error"], result.get("_status_code", 500)
+
+    return send_file(
+        result["buffer"],
+        as_attachment=True,
+        download_name=result["filename"],
+        mimetype='application/pdf'
+    )
+
+
+@api_bp.route('/download_report', methods=['POST'])
+def download_report():
+    """Generate a downloadable report from AI analysis (Works for Port & DB Vulns)"""
+    data = request.get_json() or {}
+    analysis_text = data.get('analysis', '')
+    
+    if not analysis_text:
+        return jsonify({'success': False, 'error': 'No analysis provided'}), 400
+    
+    try:
+        result = generate_ai_report(analysis_text)
+        if not result.get("success"):
+            return jsonify({'success': False, 'error': result.get('error', 'Failed to generate report')}), result.get("_status_code", 500)
+        return send_file(
+            result["buffer"],
+            mimetype=result["mimetype"],
+            as_attachment=True,
+            download_name=result["filename"],
+        )
+    
+    except Exception as e:
+        logger.exception("Error generating report")
+        return jsonify({'success': False, 'error': 'Failed to generate report', 'detail': str(e)}), 500
+
+
+@api_bp.route('/api/osint/<target>')
+def api_osint(target):
+    """
+    API endpoint for OSINT data. 
+    Keeps logic separate from the main port scanner.
+    """
+    try:
+        engine = OSINTEngine(target)
+        whois = WhoisLookup()
+        
+        results = {
+            "dns": engine.get_dns_records(),
+            "social": engine.scan_social_presence(),
+            "whois": whois.get_data(target)
+        }
+        return jsonify(results)
+    except Exception as e:
+        logger.exception("Error in OSINT endpoint")
+        return jsonify({"error": str(e)}), 500
+
+
+@api_bp.route('/api/analyze', methods=['GET'])
+def api_analyze():
+    """
+    API endpoint to analyze HTTP security headers and SSL certificate health
+    """
+    try:
+        url = request.args.get('url', '').strip()
+        
+        if not url:
+            return jsonify({'success': False, 'error': 'URL is required'}), 400
+        
+        # Ensure URL has a scheme
+        if not url.startswith(('http://', 'https://')):
+            url = f'https://{url}'
+        
+        # Security headers to check
+        required_headers = {
+            'Content-Security-Policy': 'High',
+            'Strict-Transport-Security': 'High',
+            'X-Content-Type-Options': 'Medium',
+            'X-Frame-Options': 'Medium',
+            'X-XSS-Protection': 'Medium',
+        }
+        
+        exposed_headers = {
+            'X-Powered-By': 'Low',
+            'Server': 'Low',
+            'X-AspNet-Version': 'Low',
+        }
+        
+        results = []
+        score = 50  # Start with base score
+        
+        try:
+            response = requests.get(url, timeout=10, allow_redirects=True)
+            headers = response.headers
+            
+            # Check for required security headers
+            missing_headers = 0
+            for header, severity in required_headers.items():
+                if header in headers:
+                    results.append({
+                        'header': header,
+                        'description': f'Good: {header} is configured correctly',
+                        'status': 'Present',
+                        'safe': True
+                    })
+                    score += 3
+                else:
+                    results.append({
+                        'header': header,
+                        'description': f'Missing: {header} should be configured for better security',
+                        'status': 'Missing',
+                        'safe': False
+                    })
+                    missing_headers += 1
+            
+            # Check for exposed headers (security risks)
+            for header, severity in exposed_headers.items():
+                if header in headers:
+                    results.append({
+                        'header': header,
+                        'description': f'Risk: {header} reveals server information. Consider removing it.',
+                        'status': f'Exposed: {headers.get(header, "N/A")[:30]}',
+                        'safe': False
+                    })
+                    score -= 2
+                else:
+                    results.append({
+                        'header': header,
+                        'description': f'Good: {header} is not exposed',
+                        'status': 'Not Exposed',
+                        'safe': True
+                    })
+                    score += 1
+            
+            # Check for HTTPS
+            if url.startswith('https://'):
+                try:
+                    # Attempt SSL/TLS check
+                    no_underscore_url = url.replace('https://', '').split('/')[0]
+                    sock = socket.create_connection((no_underscore_url, 443), timeout=5)
+                    context = ssl.create_default_context()
+                    with context.wrap_socket(sock, server_hostname=no_underscore_url) as ssock:
+                        cert = ssock.getpeercert()
+                        results.append({
+                            'header': 'SSL/TLS',
+                            'description': 'Good: HTTPS is enabled with valid SSL/TLS certificate',
+                            'status': 'Secure',
+                            'safe': True
+                        })
+                        score += 5
+                except Exception as e:
+                    results.append({
+                        'header': 'SSL/TLS',
+                        'description': f'Warning: SSL/TLS verification failed: {str(e)[:50]}',
+                        'status': 'Warning',
+                        'safe': False
+                    })
+            else:
+                results.append({
+                    'header': 'SSL/TLS',
+                    'description': 'Risk: Site is not using HTTPS. All data is transmitted in plaintext.',
+                    'status': 'Insecure',
+                    'safe': False
+                })
+                score -= 10
+            
+            # Normalize score between 0-100
+            score = max(0, min(100, score))
+            
+            return jsonify({
+                'success': True,
+                'score': score,
+                'results': results,
+                'url': url
+            })
+        
+        except requests.Timeout:
+            return jsonify({'success': False, 'error': 'Request timed out. Site may be unreachable.'}), 504
+        except requests.ConnectionError:
+            return jsonify({'success': False, 'error': 'Connection error. Cannot reach the target URL.'}), 502
+        except Exception as e:
+            return jsonify({'success': False, 'error': f'Failed to analyze: {str(e)[:100]}'}), 500
+    
+    except Exception as e:
+        logger.exception("Error in API analyze")
+        return jsonify({'success': False, 'error': 'Server error'}), 500
