+++ src/routes/views.py
@@ -1,4 +1,5 @@
-from flask import Blueprint, render_template, request
+from flask import Blueprint, render_template, request, redirect, url_for
+from datetime import datetime
 from extensions import latest_results
 from services.storage_service import load_history, save_message
 
@@ -8,6 +9,18 @@ views_bp = Blueprint('views', __name__)
 @views_bp.route('/', methods=['GET'])
 def landing():
     return render_template('landing_v2.html')
+
+
+@views_bp.app_context_processor
+def inject_current_year():
+    return {"current_year": datetime.now().year}
+
+@views_bp.route('/clear', methods=['POST'])
+def clear():
+    latest_results['results'] = None
+    latest_results['target'] = ''
+    latest_results['deep_scan'] = False
+    return redirect(url_for('views.dashboard'))
 
 @views_bp.route('/dashboard', methods=['GET'])
 def dashboard():
@@ -34,6 +47,15 @@ def settings_page():
     """Settings page for scanner configuration"""
     return render_template('settings.html', active_page='settings')
 
+@views_bp.route('/topology')
+def topology_page():
+    return render_template('topology.html', active_page='topology')
+
+@views_bp.route('/osint')
+def osint_page():
+    """Renders the new OSINT Reconnaissance dashboard."""
+    return render_template('osint.html', active_page='osint')
+
 @views_bp.route('/contact', methods=['GET', 'POST'])
 def contact():
     if request.method == 'POST':
@@ -58,3 +80,8 @@ def contact():
             return render_template('contact.html', error="There was an error sending your message. Please try again later.", active_page='contact')
 
     return render_template('contact.html', active_page='contact')
+
+
+@views_bp.app_errorhandler(404)
+def page_not_found(e):
+    return render_template("404.html"), 404
