+++ src/services/report_service.py
@@ -1,7 +1,27 @@
+import re
+from datetime import datetime
 from io import BytesIO
 
 from core.reporter import generate_pdf_report
+from extensions import logger
 from services.storage_service import load_history
+
+try:
+    from reportlab.lib import colors
+    from reportlab.lib.pagesizes import letter
+    from reportlab.lib.styles import ParagraphStyle, getSampleStyleSheet
+    from reportlab.lib.units import inch
+    from reportlab.platypus import Paragraph, SimpleDocTemplate, Spacer, Table, TableStyle
+    REPORTLAB_AVAILABLE = True
+except Exception:
+    REPORTLAB_AVAILABLE = False
+
+
+def _safe_filename_component(value) -> str:
+    text = str(value or "unknown")
+    text = re.sub(r'[\\/:*?"<>|]+', "-", text)
+    text = re.sub(r"\s+", "_", text).strip("._")
+    return text or "unknown"
 
 
 def generate_scan_report(scan_data) -> BytesIO:
@@ -12,7 +32,7 @@ def generate_scan_report(scan_data) -> BytesIO:
 def get_scan_by_id(scan_id) -> dict | None:
     """Fetch a scan record by id from persisted history."""
     history = load_history()
-    return next((item for item in history if item.get('id') == scan_id), None)
+    return next((item for item in history if item.get("id") == scan_id), None)
 
 
 def export_scan_report(scan_id):
@@ -22,10 +42,108 @@ def export_scan_report(scan_id):
         return {"success": False, "error": "Scan not found", "_status_code": 404}
 
     pdf_buffer = generate_scan_report(scan_data)
-    filename = f"vulnx_report_{scan_data.get('target', 'unknown')}_{scan_data.get('timestamp')}.pdf"
+    target = _safe_filename_component(scan_data.get("target", "unknown"))
+    timestamp = _safe_filename_component(scan_data.get("timestamp"))
+    filename = f"vulnx_report_{target}_{timestamp}.pdf"
 
     return {
         "success": True,
         "buffer": pdf_buffer,
         "filename": filename,
+        "mimetype": "application/pdf",
+    }
+
+
+def generate_ai_report(analysis_text):
+    """Generate downloadable AI analysis report as PDF (or TXT fallback)."""
+    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+    safe_timestamp = _safe_filename_component(timestamp)
+
+    if REPORTLAB_AVAILABLE:
+        pdf_buffer = BytesIO()
+        doc = SimpleDocTemplate(pdf_buffer, pagesize=letter)
+        story = []
+
+        styles = getSampleStyleSheet()
+        title_style = ParagraphStyle(
+            "CustomTitle",
+            parent=styles["Heading1"],
+            fontSize=24,
+            textColor=colors.HexColor("#10b981"),
+            spaceAfter=30,
+            alignment=1,
+        )
+
+        heading_style = ParagraphStyle(
+            "CustomHeading",
+            parent=styles["Heading2"],
+            fontSize=14,
+            textColor=colors.HexColor("#10b981"),
+            spaceAfter=12,
+            spaceBefore=12,
+        )
+
+        story.append(Paragraph("VulnX Security Analysis Report", title_style))
+        story.append(Spacer(1, 0.2 * inch))
+
+        metadata = [
+            ["Port", "N/A"],
+            ["Service", "N/A"],
+            ["Generated", datetime.now().strftime("%Y-%m-%d %H:%M:%S")],
+            ["Tool", "VulnX AI Scanner"],
+        ]
+
+        metadata_table = Table(metadata, colWidths=[1.5 * inch, 4 * inch])
+        metadata_table.setStyle(TableStyle([
+            ("BACKGROUND", (0, 0), (0, -1), colors.HexColor("#f0fdf4")),
+            ("TEXTCOLOR", (0, 0), (-1, -1), colors.HexColor("#0d3f2a")),
+            ("ALIGN", (0, 0), (-1, -1), "LEFT"),
+            ("FONTNAME", (0, 0), (0, -1), "Helvetica-Bold"),
+            ("FONTSIZE", (0, 0), (-1, -1), 10),
+            ("BOTTOMPADDING", (0, 0), (-1, -1), 8),
+            ("TOPPADDING", (0, 0), (-1, -1), 8),
+            ("GRID", (0, 0), (-1, -1), 1, colors.HexColor("#d1d5db")),
+        ]))
+
+        story.append(metadata_table)
+        story.append(Spacer(1, 0.3 * inch))
+        story.append(Paragraph("AI Security Analysis", heading_style))
+
+        for para in analysis_text.split("\n"):
+            if para.strip():
+                story.append(Paragraph(para.strip(), styles["Normal"]))
+                story.append(Spacer(1, 0.05 * inch))
+
+        story.append(Spacer(1, 0.3 * inch))
+        story.append(Paragraph(
+            '<font size=8 color="#999999">Report generated by VulnX AI Security Scanner | Powered by Google Gemini</font>',
+            styles["Normal"],
+        ))
+
+        doc.build(story)
+        pdf_buffer.seek(0)
+
+        return {
+            "success": True,
+            "buffer": pdf_buffer,
+            "filename": f"VulnX_Analysis_Port_{safe_timestamp}.pdf",
+            "mimetype": "application/pdf",
+        }
+
+    logger.info("reportlab not available, generating TXT report instead")
+    report_content = f"""{'='*70}
+VulnX SECURITY ANALYSIS REPORT
+{'='*70}
+
+{analysis_text}
+
+{'='*70}
+Report generated by VulnX AI Security Scanner | Powered by Google Gemini
+{'='*70}"""
+    txt_buffer = BytesIO(report_content.encode("utf-8"))
+    return {
+        "success": True,
+        "buffer": txt_buffer,
+        "filename": f"VulnX_Analysis_{safe_timestamp}.txt",
+        "mimetype": "text/plain",
     }
