+++ templates/base.html
@@ -46,14 +46,14 @@
         </a>
 
 
-      <!-- <a href="{{ url_for('topology_page') }}" data-nav-link="topology" class="nav-item {% if active_page == 'topology' %}active{% endif %}">
+      <!-- <a href="{{ url_for('views.topology_page') }}" data-nav-link="topology" class="nav-item {% if active_page == 'topology' %}active{% endif %}">
     <div class="nav-content">
         <i class="fas fa-project-diagram"></i>
         <span>Network Map</span>
     </div>
 </a> -->
 
-        <a href="{{ url_for('osint_page') }}" data-nav-link="osint"
+        <a href="{{ url_for('views.osint_page') }}" data-nav-link="osint"
            class="nav-item {% if active_page == 'osint' %}active{% endif %}"> 
            OSINT Recon
         </a>
